arreglo en validaciones,controlador  de back, variable de entorno secreto

+++ back/controllers/adminController.js
@@ -35,18 +35,18 @@ const login = async (req, res) => {
         res.json(theAdmin);
 
     } catch (error) {
-        res.status(500).json(error.message)
+        res.status(500).json({error:error.message});
     }
 };
 
 const addStudent = async (req, res) => {
     const { fullName, email, phoneNumber, program, dni, school, age, address, motive, user, pass } = req.body;
-    const avatar = req.files[0].filename;
-    const passHash = bcryptjs.hashSync(pass, 12);
     let errors = validationResult(req);
-    console.log(errors);
+    //console.log(errors);
     if (errors.isEmpty()) {
         try {
+            const avatar = req.files[0].filename;
+            const passHash = bcryptjs.hashSync(pass, 12);
             await db.query("ALTER TABLE students AUTO_INCREMENT = 1");
             await studentModel.create({
                 fullName: fullName,
@@ -64,22 +64,24 @@ const addStudent = async (req, res) => {
             });
             res.json({ message: 'Registro creado correctamente' });
         } catch (error) {
-            res.json({ message: 'Error' })
+            res.status(500).json({error:error.message})
         }
     } else {
-        res.status(500).json(errors.mapped());
-        fs.unlinkSync(path.join(__dirname,'..','..','front','src','img','students',`${avatar}`))
+        res.status(200).json(errors.mapped());
+        //console.log(req.files[0]);
+        if(typeof req.files[0] != 'undefined'){
+            const avatar = req.files[0].filename;
+            fs.unlinkSync(path.join(__dirname,'..','..','front','src','img','students',`${avatar}`));
+        }
     }
-
-
 };
 
 const getAllStudent = async (req, res) => {
     try {
-        const students = await studentModel.findAll();
-        res.json(students)
+        const students = await studentModel.findAll({attributes:['fullName','email','phoneNumber','program','avatar','dni','school','age','address','motive','user','eventId','adviserId','createdAt','updatedAt']});
+        res.json(students);
     } catch (error) {
-        res.json({ message: 'Error' })
+        res.status(500).json({error:error.message})
     }
 };
 
@@ -90,16 +92,19 @@ const getStudent = async (req, res) => {
             attributes: ['id', 'fullName', 'email', 'phoneNumber', 'program', 'avatar', 'dni', 'school', 'age', 'address', 'motive', 'user'],
             include: [{ model: adviserModel, attributes: ['id', 'fullName', 'email', 'phoneNumber'] }, { model: eventModel, attributes: ['id', 'name', 'date', 'time', 'detail', 'duration', 'adviser_event_id'] }]
         })
-        res.json(student[0])
+        if(!student[0]){
+            return res.status(404).json({message:'404 user not found'});
+        }
+        res.json(student[0]);
     } catch (error) {
-        res.status(500).json({ message: error.message })
+        res.status(500).json({error:error.message})
     }
 };
 
 const logOut = (req = request, res) => {
 
     req.session.destroy((err) => {
-        res.clearCookie('userId').send('cleared cookie');//encontrado en stack overflow,no lo se explicar
+        res.clearCookie('userId').send('cleared cookie');//por alguna razon al agregar el send si funciona
     });
 
 };
